app/requirements.txt
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-CORS==2.1.2
pymongo==3.2.2
jsonpickle==0.9.3
itsdangerous==0.24
setuptools==25.1.1
html5lib==1.0b8
pytest==2.9.1
pytest-flask==0.10.0
pytest-cov==2.2.1
pytest-xdist==1.14
python-slugify==1.2.0
bleach==1.4.2
freezegun==0.3.7
passlib==1.6.5
honcho==0.7.1
mongomock==3.3.0
gunicorn==19.5.0
gevent==1.1.1
sentinels==0.0.6
mock==1.0.1
requests==2.20.0
requests-mock==0.7.0
py==1.4.32
